+++ docs/conf.py
@@ -10,7 +10,7 @@ sys.path.insert(0, str(SRC))
 
 project = "shmpipeline"
 author = "Jacob Taylor"
-release = "1.0.2"
+release = "1.0.3"
 version = release
 
 extensions = [